Revert split.
