+++ src/services/fetch.rs
@@ -3,20 +3,57 @@
 use super::Task;
 use crate::callback::Callback;
 use crate::format::{Binary, Format, Text};
-use serde::Serialize;
-use std::collections::HashMap;
+use cfg_if::cfg_if;
+use cfg_match::cfg_match;
+use failure::Fail;
 use std::fmt;
-use stdweb::serde::Serde;
-use stdweb::unstable::{TryFrom, TryInto};
-use stdweb::web::ArrayBuffer;
-use stdweb::{JsSerialize, Value};
-#[allow(unused_imports)]
-use stdweb::{_js_impl, js};
-use thiserror::Error;
+cfg_if! {
+    if #[cfg(feature = "std_web")] {
+        use serde::Serialize;
+        use std::collections::HashMap;
+        use stdweb::serde::Serde;
+        use stdweb::unstable::{TryFrom, TryInto};
+        use stdweb::web::ArrayBuffer;
+        use stdweb::{JsSerialize, Value};
+        #[allow(unused_imports)]
+        use stdweb::{_js_impl, js};
+    } else if #[cfg(feature = "web_sys")] {
+        use js_sys::{Array, Promise, Reflect, Uint8Array};
+        use std::rc::Rc;
+        use std::sync::atomic::{AtomicBool, Ordering};
+        use std::sync::mpsc::{self, Receiver};
+        use wasm_bindgen::{closure::Closure, JsValue};
+        use web_sys::{
+            AbortController, Headers, Request as WebRequest, RequestInit, Response as WebResponse,
+        };
+        pub use web_sys::{
+            RequestCache as Cache, RequestCredentials as Credentials, RequestMode as Mode,
+            RequestRedirect as Redirect, Window, WorkerGlobalScope,
+        };
+    }
+}
 
 pub use http::{HeaderMap, Method, Request, Response, StatusCode, Uri};
 
+#[cfg(feature = "web_sys")]
+struct ArrayBuffer(Uint8Array);
+
+#[cfg(feature = "web_sys")]
+impl From<ArrayBuffer> for Vec<u8> {
+    fn from(from: ArrayBuffer) -> Self {
+        from.0.to_vec()
+    }
+}
+
+#[cfg(feature = "web_sys")]
+impl From<JsValue> for ArrayBuffer {
+    fn from(from: JsValue) -> Self {
+        ArrayBuffer(Uint8Array::new_with_byte_offset(&from, 0))
+    }
+}
+
 /// Type to set cache for fetch.
+#[cfg(feature = "std_web")]
 #[derive(Serialize, Debug)]
 #[serde(rename_all = "kebab-case")]
 pub enum Cache {
@@ -36,6 +73,7 @@ pub enum Cache {
 }
 
 /// Type to set credentials for fetch.
+#[cfg(feature = "std_web")]
 #[derive(Serialize, Debug)]
 #[serde(rename_all = "kebab-case")]
 pub enum Credentials {
@@ -48,6 +86,7 @@ pub enum Credentials {
 }
 
 /// Type to set mode for fetch.
+#[cfg(feature = "std_web")]
 #[derive(Serialize, Debug)]
 #[serde(rename_all = "kebab-case")]
 pub enum Mode {
@@ -60,6 +99,7 @@ pub enum Mode {
 }
 
 /// Type to set redirect behaviour for fetch.
+#[cfg(feature = "std_web")]
 #[derive(Serialize, Debug)]
 #[serde(rename_all = "kebab-case")]
 pub enum Redirect {
@@ -73,20 +113,46 @@ pub enum Redirect {
 
 /// Init options for `fetch()` function call.
 /// https://developer.mozilla.org/en-US/docs/Web/API/WindowOrWorkerGlobalScope/fetch
-#[derive(Serialize, Default, Debug)]
+#[cfg_attr(feature = "std_web", derive(Serialize))]
+#[derive(Default, Debug)]
 pub struct FetchOptions {
     /// Cache of a fetch request.
-    #[serde(skip_serializing_if = "Option::is_none")]
+    #[cfg_attr(feature = "std_web", serde(skip_serializing_if = "Option::is_none"))]
     pub cache: Option<Cache>,
     /// Credentials of a fetch request.
-    #[serde(skip_serializing_if = "Option::is_none")]
+    #[cfg_attr(feature = "std_web", serde(skip_serializing_if = "Option::is_none"))]
     pub credentials: Option<Credentials>,
     /// Redirect behaviour of a fetch request.
-    #[serde(skip_serializing_if = "Option::is_none")]
+    #[cfg_attr(feature = "std_web", serde(skip_serializing_if = "Option::is_none"))]
     pub redirect: Option<Redirect>,
     /// Request mode of a fetch request.
-    #[serde(skip_serializing_if = "Option::is_none")]
+    #[cfg_attr(feature = "std_web", serde(skip_serializing_if = "Option::is_none"))]
     pub mode: Option<Mode>,
+}
+
+#[cfg(feature = "web_sys")]
+impl Into<RequestInit> for FetchOptions {
+    fn into(self) -> RequestInit {
+        let mut init = RequestInit::new();
+
+        if let Some(cache) = self.cache {
+            init.cache(cache);
+        }
+
+        if let Some(credentials) = self.credentials {
+            init.credentials(credentials);
+        }
+
+        if let Some(redirect) = self.redirect {
+            init.redirect(redirect);
+        }
+
+        if let Some(mode) = self.mode {
+            init.mode(mode);
+        }
+
+        init
+    }
 }
 
 /// Represents errors of a fetch service.
@@ -96,9 +162,21 @@ enum FetchError {
     FailedResponse,
 }
 
+#[cfg(feature = "web_sys")]
+#[derive(Debug)]
+struct Handle {
+    active: Rc<AtomicBool>,
+    callbacks: Receiver<Closure<dyn FnMut(JsValue)>>,
+    abort_controller: Option<AbortController>,
+    promise: Promise,
+}
+
 /// A handle to control sent requests. Can be canceled with a `Task::cancel` call.
 #[must_use]
-pub struct FetchTask(Option<Value>);
+pub struct FetchTask(
+    #[cfg(feature = "std_web")] Option<Value>,
+    #[cfg(feature = "web_sys")] Option<Handle>,
+);
 
 impl fmt::Debug for FetchTask {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
@@ -223,7 +301,19 @@ impl FetchService {
         IN: Into<Text>,
         OUT: From<Text>,
     {
-        fetch_impl::<IN, OUT, String, String>(false, request, None, callback)
+        cfg_match! {
+            feature = "std_web" => fetch_impl::<IN, OUT, String, String>(false, request, None, callback),
+            feature = "web_sys" => ({
+                fetch_impl::<IN, OUT, String, String, _, _>(
+                    false,
+                    request,
+                    None,
+                    callback,
+                    Into::into,
+                    |v| v.as_string().unwrap(),
+                )
+            }),
+        }
     }
 
     /// `fetch` with provided `FetchOptions` object.
@@ -266,7 +356,19 @@ impl FetchService {
         IN: Into<Text>,
         OUT: From<Text>,
     {
-        fetch_impl::<IN, OUT, String, String>(false, request, Some(options), callback)
+        cfg_match! {
+            feature = "std_web" => fetch_impl::<IN, OUT, String, String>(false, request, Some(options), callback),
+            feature = "web_sys" => ({
+                fetch_impl::<IN, OUT, String, String, _, _>(
+                    false,
+                    request,
+                    Some(options),
+                    callback,
+                    Into::into,
+                    |v| v.as_string().unwrap(),
+                )
+            }),
+        }
     }
 
     /// Fetch the data in binary format.
@@ -279,7 +381,19 @@ impl FetchService {
         IN: Into<Binary>,
         OUT: From<Binary>,
     {
-        fetch_impl::<IN, OUT, Vec<u8>, ArrayBuffer>(true, request, None, callback)
+        cfg_match! {
+            feature = "std_web" => fetch_impl::<IN, OUT, Vec<u8>, ArrayBuffer>(true, request, None, callback),
+            feature = "web_sys" => ({
+                fetch_impl::<IN, OUT, Vec<u8>, ArrayBuffer, _, _>(
+                    true,
+                    request,
+                    None,
+                    callback,
+                    |v| Uint8Array::from(v.as_slice()).into(),
+                    From::from,
+                )
+            }),
+        }
     }
 
     /// Fetch the data in binary format.
@@ -293,39 +407,64 @@ impl FetchService {
         IN: Into<Binary>,
         OUT: From<Binary>,
     {
-        fetch_impl::<IN, OUT, Vec<u8>, ArrayBuffer>(true, request, Some(options), callback)
+        cfg_match! {
+            feature = "std_web" => fetch_impl::<IN, OUT, Vec<u8>, ArrayBuffer>(true, request, Some(options), callback),
+            feature = "web_sys" => ({
+                fetch_impl::<IN, OUT, Vec<u8>, ArrayBuffer, _, _>(
+                    true,
+                    request,
+                    Some(options),
+                    callback,
+                    |v| Uint8Array::from(v.as_slice()).into(),
+                    From::from,
+                )
+            }),
+        }
     }
 }
 
-fn fetch_impl<IN, OUT: 'static, T, X>(
+fn fetch_impl<
+    IN,
+    OUT: 'static,
+    #[cfg(feature = "std_web")] T: JsSerialize,
+    #[cfg(feature = "web_sys")] T,
+    #[cfg(feature = "std_web")] X: TryFrom<Value> + Into<T>,
+    #[cfg(feature = "web_sys")] X: Into<T>,
+    #[cfg(feature = "web_sys")] IC: Fn(T) -> JsValue,
+    #[cfg(feature = "web_sys")] FC: 'static + Fn(JsValue) -> X,
+>(
     binary: bool,
     request: Request<IN>,
     options: Option<FetchOptions>,
     callback: Callback<Response<OUT>>,
+    #[cfg(feature = "web_sys")] into_conversion: IC,
+    #[cfg(feature = "web_sys")] from_conversion: FC,
 ) -> FetchTask
 where
     IN: Into<Format<T>>,
     OUT: From<Format<T>>,
-    T: JsSerialize,
-    X: TryFrom<Value> + Into<T>,
 {
     // Consume request as parts and body.
     let (parts, body) = request.into_parts();
 
     // Map headers into a Js serializable HashMap.
-    let header_map: HashMap<&str, &str> = parts
-        .headers
-        .iter()
-        .map(|(k, v)| {
-            (
-                k.as_str(),
-                v.to_str().unwrap_or_else(|_| {
-                    panic!("Unparsable request header {}: {:?}", k.as_str(), v)
-                }),
-            )
-        })
-        .collect();
-
+    let header_map = parts.headers.iter().map(|(k, v)| {
+        (
+            k.as_str(),
+            v.to_str()
+                .unwrap_or_else(|_| panic!("Unparsable request header {}: {:?}", k.as_str(), v)),
+        )
+    });
+    let header_map = cfg_match! {
+        feature = "std_web" => header_map.collect::<HashMap<&str, &str>>(),
+        feature = "web_sys" => ({
+            let headers = Headers::new().unwrap();
+            for (k, v) in header_map {
+                headers.append(k, v).unwrap();
+            }
+            headers
+        }),
+    };
     // Formats URI.
     let uri = format!("{}", parts.uri);
     let method = parts.method.as_str();
@@ -334,14 +473,37 @@ where
     // Prepare the response callback.
     // Notice that the callback signature must match the call from the javascript
     // side. There is no static check at this point.
-    let callback = move |success: bool, status: u16, headers: HashMap<String, String>, data: X| {
+    let callback = move |#[cfg(feature = "std_web")] success: bool,
+                         #[cfg(feature = "web_sys")] data: Option<X>,
+                         status: u16,
+                         #[cfg(feature = "std_web")] headers: HashMap<String, String>,
+                         #[cfg(feature = "web_sys")] headers: Headers,
+                         #[cfg(feature = "std_web")] data: X| {
         let mut response_builder = Response::builder().status(status);
-        for (key, values) in headers {
-            response_builder = response_builder.header(key.as_str(), values.as_str());
+        // convert `headers` to `Iterator<Item = (String, String)>`
+        let headers = cfg_match! {
+            feature = "std_web" => headers.into_iter(),
+            feature = "web_sys" => ({
+                js_sys::try_iter(&headers)
+                    .unwrap()
+                    .unwrap()
+                    .map(Result::unwrap)
+                    .map(|entry| {
+                        let entry = Array::from(&entry);
+                        let key = entry.get(0);
+                        let value = entry.get(1);
+                        (key.as_string().unwrap(), value.as_string().unwrap())
+                    })
+            }),
+        };
+        for (key, value) in headers {
+            response_builder = response_builder.header(key.as_str(), value.as_str());
         }
 
         // Deserialize and wrap response data into a Text object.
-        let data = if success {
+        #[cfg(feature = "std_web")]
+        let data = Some(data).filter(|_| success);
+        let data = if let Some(data) = data {
             Ok(data.into())
         } else {
             Err(FetchError::FailedResponse.into())
@@ -352,57 +514,134 @@ where
     };
 
     #[allow(clippy::too_many_arguments)]
-    let handle = js! {
-        var body = @{body};
-        if (@{binary} && body != null) {
-            body = Uint8Array.from(body);
-        }
-        var data = {
-            method: @{method},
-            body: body,
-            headers: @{header_map},
-        };
-        var request = new Request(@{uri}, data);
-        var callback = @{callback};
-        var abortController = AbortController ? new AbortController() : null;
-        var handle = {
-            active: true,
-            callback,
-            abortController,
-        };
-        var init = @{Serde(options)} || {};
-        if (abortController && !("signal" in init)) {
-            init.signal = abortController.signal;
-        }
-        fetch(request, init).then(function(response) {
-            var promise = (@{binary}) ? response.arrayBuffer() : response.text();
-            var status = response.status;
-            var headers = {};
-            response.headers.forEach(function(value, key) {
-                headers[key] = value;
-            });
-            promise.then(function(data) {
-                if (handle.active == true) {
-                    handle.active = false;
-                    callback(true, status, headers, data);
-                    callback.drop();
-                }
-            }).catch(function(err) {
-                if (handle.active == true) {
-                    handle.active = false;
-                    callback(false, status, headers, data);
-                    callback.drop();
-                }
-            });
-        }).catch(function(e) {
-            if (handle.active == true) {
-                var data = (@{binary}) ? new ArrayBuffer() : "";
-                handle.active = false;
-                callback(false, 408, {}, data);
-                callback.drop();
+    let handle = cfg_match! {
+        feature = "std_web" => js! {
+            var body = @{body};
+            if (@{binary} && body != null) {
+                body = Uint8Array.from(body);
             }
-        });
-        return handle;
+            var data = {
+                method: @{method},
+                body: body,
+                headers: @{header_map},
+            };
+            var request = new Request(@{uri}, data);
+            var callback = @{callback};
+            var abortController = AbortController ? new AbortController() : null;
+            var handle = {
+                active: true,
+                callback,
+                abortController,
+            };
+            var init = @{Serde(options)} || {};
+            if (abortController && !("signal" in init)) {
+                init.signal = abortController.signal;
+            }
+            fetch(request, init).then(function(response) {
+                var promise = (@{binary}) ? response.arrayBuffer() : response.text();
+                var status = response.status;
+                var headers = {};
+                response.headers.forEach(function(value, key) {
+                    headers[key] = value;
+                });
+                promise.then(function(data) {
+                    if (handle.active == true) {
+                        handle.active = false;
+                        callback(true, status, headers, data);
+                        callback.drop();
+                    }
+                }).catch(function(err) {
+                    if (handle.active == true) {
+                        handle.active = false;
+                        callback(false, status, headers, data);
+                        callback.drop();
+                    }
+                });
+            }).catch(function(e) {
+                if (handle.active == true) {
+                    var data = (@{binary}) ? new ArrayBuffer() : "";
+                    handle.active = false;
+                    callback(false, 408, {}, data);
+                    callback.drop();
+                }
+            });
+            return handle;
+        },
+        feature = "web_sys" => ({
+            let mut data = RequestInit::new();
+            data.method(method);
+            data.body(body.map(into_conversion).as_ref());
+            data.headers(&header_map);
+            let request = WebRequest::new_with_str_and_init(&uri, &data).unwrap();
+            let active = Rc::new(AtomicBool::new(true));
+            let (sender, receiver) = mpsc::channel();
+            let active_outer_clone = Rc::clone(&active);
+            let callback_outer_clone = callback.clone();
+            let sender_clone = sender.clone();
+            let closure_then = move |response: JsValue| {
+                let response = WebResponse::from(response);
+                let promise = if binary {
+                    response.array_buffer()
+                } else {
+                    response.text()
+                }
+                .unwrap();
+                let status = response.status();
+                let headers = response.headers();
+                let active_clone = Rc::clone(&active_outer_clone);
+                let callback_clone = callback_outer_clone.clone();
+                let headers_clone = headers.clone();
+                let closure_then = move |data: JsValue| {
+                    let data = from_conversion(data);
+                    if active_clone.compare_and_swap(true, false, Ordering::SeqCst) {
+                        callback_clone(Some(data), status, headers_clone);
+                    }
+                };
+                let closure_then = Closure::once(closure_then);
+                let closure_catch = move |_| {
+                    if active_outer_clone.compare_and_swap(true, false, Ordering::SeqCst) {
+                        callback_outer_clone(None, status, headers);
+                    }
+                };
+                let closure_catch = Closure::once(closure_catch);
+                #[allow(unused_must_use)]
+                {
+                    promise.then(&closure_then).catch(&closure_catch);
+                }
+                sender_clone.send(closure_then).unwrap();
+                sender_clone.send(closure_catch).unwrap();
+            };
+            let closure_then = Closure::once(closure_then);
+            let active_clone = Rc::clone(&active);
+            let closure_catch = move |_| {
+                if active_clone.compare_and_swap(true, false, Ordering::SeqCst) {
+                    callback(None, 408, Headers::new().unwrap());
+                }
+            };
+            let closure_catch = Closure::wrap(Box::new(closure_catch) as Box<dyn FnMut(JsValue)>);
+            let abort_controller = AbortController::new().ok();
+            let mut init = options.map_or_else(RequestInit::new, Into::into);
+            if let Some(abort_controller) = &abort_controller {
+                init.signal(Some(&abort_controller.signal()));
+            }
+            let global: JsValue = js_sys::global().into();
+            let promise = if Reflect::has(&global, &String::from("Window").into()).unwrap() {
+                Window::from(global).fetch_with_request_and_init(&request, &init)
+            } else if Reflect::has(&global, &String::from("WorkerGlobalScope").into()).unwrap() {
+                WorkerGlobalScope::from(global).fetch_with_request_and_init(&request, &init)
+            } else {
+                panic!("failed to get global context")
+            };
+            let promise = promise.then(&closure_then).catch(&closure_catch);
+            sender.send(closure_then).unwrap();
+            sender.send(closure_catch).unwrap();
+            Handle {
+                active,
+                callbacks: receiver,
+                abort_controller,
+                promise,
+            }
+        }),
     };
     FetchTask(Some(handle))
 }
@@ -410,12 +649,25 @@ where
 impl Task for FetchTask {
     fn is_active(&self) -> bool {
         if let Some(ref task) = self.0 {
-            let result = js! {
-                var the_task = @{task};
-                return the_task.active &&
-                        (!the_task.abortController || !the_task.abortController.signal.aborted);
-            };
-            result.try_into().unwrap_or(false)
+            cfg_match! {
+                feature = "std_web" => ({
+                    let result = js! {
+                        var the_task = @{task};
+                        return the_task.active &&
+                                (!the_task.abortController || !the_task.abortController.signal.aborted);
+                    };
+                    result.try_into().unwrap_or(false)
+                }),
+                feature = "web_sys" => ({
+                    task.active.load(Ordering::SeqCst)
+                        && task
+                            .abort_controller
+                            .as_ref()
+                            .map(|abort_controller| abort_controller.signal().aborted())
+                            .filter(|value| *value)
+                            .is_none()
+                }),
+            }
         } else {
             false
         }
@@ -428,14 +680,32 @@ impl Task for FetchTask {
             .0
             .take()
             .expect("tried to cancel request fetching twice");
-        js! {  @(no_return)
-            var handle = @{handle};
-            handle.active = false;
-            handle.callback.drop();
-            if (handle.abortController) {
-                handle.abortController.abort();
-            }
-        }
+        cfg_match! {
+            feature = "std_web" => ({
+                js! {  @(no_return)
+                    var handle = @{handle};
+                    handle.active = false;
+                    handle.callback.drop();
+                    if (handle.abortController) {
+                        handle.abortController.abort();
+                    }
+                };
+            }),
+            feature = "web_sys" => ({
+                thread_local! {
+                    static CATCH: Closure<dyn FnMut(JsValue)> = Closure::wrap(Box::new(|_| ()) as Box<dyn FnMut(JsValue)>);
+                }
+                handle.active.store(false, Ordering::SeqCst);
+                #[allow(unused_must_use)]
+                {
+                    CATCH.with(|c| handle.promise.catch(&c));
+                }
+                if let Some(abort_controller) = handle.abort_controller {
+                    abort_controller.abort();
+                }
+                handle.callbacks.try_iter().for_each(drop);
+            }),
+        };
     }
 }
